Forty-three signatures that stop saying "anything can happen"
S112: every action handler declared "throws Exception", so every caller had to
catch everything and no reader could tell which handlers can actually fail in a
checked way. They cannot, with one exception — POI writes files, and that is an
IOException. Forty-three signatures narrowed to it.

The one handler that really did throw something else is the interesting one:
SparklineHandler builds an x14 extension as XML text and parses it, which throws
XmlException. That is now named where it happens, and wrapped as an
IllegalStateException with the reason — the XML is built in that class from
values it has already validated, so a parse failure there is a bug in this code
rather than anything the caller did.

S8700 and S8688 in AuthController, which is where the .now() family actually
matters: the refresh cookie's max-age was a Duration between two bare local
times, and a duration between local times is off by an hour across a
daylight-saving change. Both ends are anchored to a zone now. The rest of the
.now() calls in the codebase are timestamps rather than arithmetic, and are left
for the Clock question.

S5853, second pass: the chains the first pass could not see, because they were
multi-line or had the sentence explaining them sitting in between. Fourteen
joined, with each comment kept above the calls it is about — inside the chain,
which reads better than it sounds.

Plus S5838 and S1612 in the round-trip test, an int multiplication widened to
double after the fact, and a cast to long of something already being widened.

1006 tests green.

+++ sheetsmith-java/src/main/java/com/ap0stole/sheetsmith/services/excel/actions/chart/BarChartStrategy.java
@@ -92,7 +92,7 @@ public class BarChartStrategy implements ChartStrategy {
             byte[] rgb = (i < TOP_20_COLORS.length) ? TOP_20_COLORS[i] : generateColor(i, pointCount);
 
             CTDPt dPt = ctBarSer.addNewDPt();
-            dPt.addNewIdx().setVal((long) i);
+            dPt.addNewIdx().setVal(i);
 
             CTShapeProperties spPr = dPt.addNewSpPr();
             spPr.addNewSolidFill().addNewSrgbClr().setVal(rgb);
